move chat handler funcs
