+++ internal/service/chat/handlers.go
@@ -4,8 +4,53 @@ import (
 	"log"
 	"net/http"
 
+	"github.com/brianaung/rtm/internal/auth"
+	"github.com/brianaung/rtm/ui"
 	"github.com/go-chi/chi/v5"
 )
+
+func (s *service) handleDashboard(w http.ResponseWriter, r *http.Request) {
+	userData := r.Context().Value("user").(*auth.UserContext)
+	w.WriteHeader(http.StatusFound)
+
+	roomids := make([]string, 0)
+	for k := range s.hubs {
+		roomids = append(roomids, k)
+	}
+	pagedata := struct {
+		User    *auth.UserContext
+		Roomids []string
+	}{
+		User:    userData,
+		Roomids: roomids,
+	}
+
+	ui.RenderPage(w, pagedata, "dashboard")
+}
+
+func (s *service) handleCreateRoom(w http.ResponseWriter, r *http.Request) {
+	roomid := r.FormValue("roomid")
+	if _, ok := s.hubs[roomid]; ok {
+		w.Write([]byte("Room already exists"))
+		return
+	}
+
+	h := newHub()
+	s.hubs[roomid] = h
+	go h.run()
+
+	w.Header().Set("HX-Redirect", "/dashboard/room/"+roomid)
+	w.WriteHeader(http.StatusOK)
+}
+
+func (s *service) handleJoinRoom(w http.ResponseWriter, r *http.Request) {
+	roomid := chi.URLParam(r, "roomid")
+	if _, ok := s.hubs[roomid]; !ok {
+		// todo: error handle
+		return
+	}
+	ui.RenderPage(w, struct{ RoomId string }{RoomId: roomid}, "chatroom")
+}
 
 func (s *service) handleServeWs(w http.ResponseWriter, r *http.Request) {
 	roomid := chi.URLParam(r, "roomid")
@@ -29,3 +74,4 @@ func (s *service) handleServeWs(w http.ResponseWriter, r *http.Request) {
 	go client.writePump()
 	go client.readPump()
 }
+
