+++ internal/service/user/handlers.go
@@ -3,7 +3,6 @@ package user
 import (
 	"net/http"
 
-	"github.com/brianaung/rtm/internal/auth"
 	"github.com/brianaung/rtm/ui"
 )
 
@@ -82,9 +81,3 @@ func (s *service) handleLogout(w http.ResponseWriter, r *http.Request) {
 }
 
 /* ================================================ */
-
-func (s *service) handleDashboard(w http.ResponseWriter, r *http.Request) {
-	userData := r.Context().Value("user").(*auth.UserContext)
-	w.WriteHeader(http.StatusFound)
-	ui.RenderPage(w, userData, "dashboard")
-}
